Fix for NYT returning an article with no images
When the NYT api returns an article with no images, it returns
"media": "",
instead of
"media":[],
This makes gson think it's a string when it's expecting an array, causing the app to crash.
I added a line that replaces the incorrect json with the correct one, and uses a placeholder image if none can be found.